changed ns, npm, and pod
