+++ npm/pod.go
@@ -3,8 +3,6 @@
 package npm
 
 import (
-	"strings"
-
 	"github.com/Azure/azure-container-networking/log"
 	"github.com/Azure/azure-container-networking/npm/util"
 
@@ -20,60 +18,6 @@ func isValidPod(podObj *corev1.Pod) bool {
 
 func isSystemPod(podObj *corev1.Pod) bool {
 	return podObj.ObjectMeta.Namespace == util.KubeSystemFlag
-}
-
-// AddPod handles adding pod ip to its label's ipset.
-func (npMgr *NetworkPolicyManager) AddPod(podObj *corev1.Pod) error {
-	npMgr.Lock()
-	defer npMgr.Unlock()
-
-	if !isValidPod(podObj) {
-		return nil
-	}
-
-	var err error
-
-	podNs := podObj.ObjectMeta.Namespace
-	podName := podObj.ObjectMeta.Name
-	podNodeName := podObj.Spec.NodeName
-	podLabels := podObj.ObjectMeta.Labels
-	podIP := podObj.Status.PodIP
-	log.Printf("POD CREATING: [%s/%s/%s%+v%s]", podNs, podName, podNodeName, podLabels, podIP)
-
-	// Add the pod to ipset
-	ipsMgr := npMgr.nsMap[util.KubeAllNamespacesFlag].ipsMgr
-	// Add the pod to its namespace's ipset.
-	log.Printf("Adding pod %s to ipset %s", podIP, podNs)
-	if err = ipsMgr.AddToSet(podNs, podIP); err != nil {
-		log.Errorf("Error: failed to add pod to namespace ipset.")
-		return err
-	}
-
-	// Add the pod to its label's ipset.
-	var labelKeys []string
-	for podLabelKey, podLabelVal := range podLabels {
-		//Ignore pod-template-hash label.
-		if strings.Contains(podLabelKey, util.KubePodTemplateHashFlag) {
-			continue
-		}
-
-		labelKey := util.KubeAllNamespacesFlag + "-" + podLabelKey + ":" + podLabelVal
-		log.Printf("Adding pod %s to ipset %s", podIP, labelKey)
-		if err = ipsMgr.AddToSet(labelKey, podIP); err != nil {
-			log.Errorf("Error: failed to add pod to label ipset.")
-			return err
-		}
-		labelKeys = append(labelKeys, labelKey)
-	}
-
-	ns, err := newNs(podNs)
-	if err != nil {
-		log.Errorf("Error: failed to create namespace %s", podNs)
-		return err
-	}
-	npMgr.nsMap[podNs] = ns
-
-	return nil
 }
 
 // UpdatePod handles updating pod ip in its label's ipset.
@@ -107,48 +51,6 @@ func (npMgr *NetworkPolicyManager) UpdatePod(oldPodObj, newPodObj *corev1.Pod) e
 
 	if newPodObj.ObjectMeta.DeletionTimestamp == nil && newPodObj.ObjectMeta.DeletionGracePeriodSeconds == nil {
 		if err = npMgr.AddPod(newPodObj); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-// DeletePod handles deleting pod from its label's ipset.
-func (npMgr *NetworkPolicyManager) DeletePod(podObj *corev1.Pod) error {
-	npMgr.Lock()
-	defer npMgr.Unlock()
-
-	if !isValidPod(podObj) {
-		return nil
-	}
-
-	var err error
-
-	podNs := podObj.ObjectMeta.Namespace
-	podName := podObj.ObjectMeta.Name
-	podNodeName := podObj.Spec.NodeName
-	podLabels := podObj.ObjectMeta.Labels
-	podIP := podObj.Status.PodIP
-	log.Printf("POD DELETING: [%s/%s/%s%+v%s]", podNs, podName, podNodeName, podLabels, podIP)
-
-	// Delete pod from ipset
-	ipsMgr := npMgr.nsMap[util.KubeAllNamespacesFlag].ipsMgr
-	// Delete the pod from its namespace's ipset.
-	if err = ipsMgr.DeleteFromSet(podNs, podIP); err != nil {
-		log.Errorf("Error: failed to delete pod from namespace ipset.")
-		return err
-	}
-	// Delete the pod from its label's ipset.
-	for podLabelKey, podLabelVal := range podLabels {
-		//Ignore pod-template-hash label.
-		if strings.Contains(podLabelKey, "pod-template-hash") {
-			continue
-		}
-
-		labelKey := util.KubeAllNamespacesFlag + "-" + podLabelKey + ":" + podLabelVal
-		if err = ipsMgr.DeleteFromSet(labelKey, podIP); err != nil {
-			log.Errorf("Error: failed to delete pod from label ipset.")
 			return err
 		}
 	}
